new mpts-demux version 2: fix AUD handling and timing assignment for output slices packaged along delimiter boundaries.
AUD is always terminating a group of picture slices (may also be IDRs eg ULL encoding where multiple keyframe slices are output for one frame - here no special handling is needed for that). The used packet timing should be the one of either first slice in group. Specifically timing of AUD NAL if properly muxed should be equal to that and all other slices, but just to make sure we rely on the first NAL anyway.

Noticeable is how the overall logic to all this is way more comprehensible as streamlined and sensible itfp in comparison to the spagethi-special-casing when using mux-js s partly erratic output in terms of timing etc.